+++ src/main/java/com/fabiogouw/bank/adapters/actors/messages/DepositResponse.java
@@ -1,9 +1,11 @@
 package com.fabiogouw.bank.adapters.actors.messages;
 
+import java.math.BigDecimal;
+
 public class DepositResponse extends OperationResponse {
     private static final long serialVersionUID = -5136902613153736547L;
 
-    public DepositResponse(String correlationId, double currentBalance, Boolean success) {
+    public DepositResponse(String correlationId, BigDecimal currentBalance, Boolean success) {
         super(correlationId, currentBalance, success);
     }
 }
