+++ src/main/java/com/fabiogouw/bank/adapters/dtos/AccountBalanceResponse.java
@@ -1,14 +1,16 @@
 package com.fabiogouw.bank.adapters.dtos;
 
+import java.math.BigDecimal;
+
 public class AccountBalanceResponse {
     private String _accountId;
-    private double _balance;
+    private BigDecimal _balance;
 
     public AccountBalanceResponse() {
 
     }
 
-    public AccountBalanceResponse(String accountId, double balance) {
+    public AccountBalanceResponse(String accountId, BigDecimal balance) {
         _accountId = accountId;
         _balance = balance;
     }
@@ -16,13 +18,13 @@ public class AccountBalanceResponse {
     public String getAccountId() {
         return _accountId;
     }
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return _balance;
     }
     public void setAccountId(String accountId){
         _accountId = accountId;
     }
-    public void setBalance(double amount) {
+    public void setBalance(BigDecimal amount) {
         _balance = amount;
     }
 }
