+++ src/main/java/com/fabiogouw/bank/adapters/dtos/OperationResponse.java
@@ -1,16 +1,18 @@
 package com.fabiogouw.bank.adapters.dtos;
 
+import java.math.BigDecimal;
+
 public class OperationResponse {
     private String _correlationId;
-    private double _amount;
-    private double _currentBalance;
+    private BigDecimal _amount;
+    private BigDecimal _currentBalance;
     private Boolean _success;
 
     public OperationResponse() {
 
     }
 
-    public OperationResponse(String correlationId, double amount, double currentBalance, Boolean success) {
+    public OperationResponse(String correlationId, BigDecimal amount, BigDecimal currentBalance, Boolean success) {
         _correlationId = correlationId;
         _amount = amount;
         _currentBalance = currentBalance;
@@ -20,10 +22,10 @@ public class OperationResponse {
     public String getCorrelationId() {
         return _correlationId;
     }
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return _amount;
     }
-    public double getCurrentBalance() {
+    public BigDecimal getCurrentBalance() {
         return _currentBalance;
     }
     public Boolean getSuccess() {
@@ -32,10 +34,10 @@ public class OperationResponse {
     public void setCorrelationId(String correlationId){
         _correlationId = correlationId;
     }
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         _amount = amount;
     }
-    public void setCurrentBalance(double currentBalance) {
+    public void setCurrentBalance(BigDecimal currentBalance) {
         _currentBalance = currentBalance;
     }    
     public void setSuccess(Boolean success) {
